Added separate event to exit screen with result
